moved variable declarations to source files for I2S and EspRingBuffer
I2SIO now instantiates both in and output I2S, although, it is untested

should probably branch all of these unfinished things and merge into master later
instead of adding all this half working or nonfunctional nonsense to master

+++ dishrec_esp32/src/RingBuffer.cpp
@@ -41,7 +41,7 @@ std::vector<I> floatToInt(std::vector<F> values)
     size_t length(values.size());
     std::vector<I> converted;
     converted.reserve(length);
-    for (int i(0); i < length; ++i)
+    for (size_t i(0); i < length; ++i)
     {
         converted.emplace_back(floatToInt(values[i]));
     }
@@ -66,7 +66,7 @@ bufferLength(bufferSize)
     {
         this->ring.emplace_back(std::vector<T>());
         this->ring[i].reserve(this->bufferLength);
-        for (int j(0); j < this->bufferLength; ++j)
+        for (uint32_t j(0); j < this->bufferLength; ++j)
         {
             this->ring[i].emplace_back(this->_zero);
         }
